+++ common/src/main/java/com/liskovsoft/smartyoutubetv2/common/prefs/GeneralData.java
@@ -86,6 +86,7 @@ public class GeneralData implements ProfileChangeListener {
     private int mLocalDriveBackupFreqDays;
     private List<Video> mOldPinnedItems;
     private boolean mIsRemapSToSpeedToggleEnabled;
+    private boolean mIsConscryptEnabled;
     private final Runnable mPersistStateInt = this::persistStateInt;
 
     private GeneralData(Context context) {
@@ -575,6 +576,15 @@ public class GeneralData implements ProfileChangeListener {
         persistState();
     }
 
+    public boolean isConscryptEnabled() {
+        return mIsConscryptEnabled;
+    }
+
+    public void setConscryptEnabled(boolean enable) {
+        mIsConscryptEnabled = enable;
+        persistState();
+    }
+
     /**
      * Fixed ConcurrentModificationException after onProfileChanged()<br/>
      * Happened inside cleanupPinnedItems()
@@ -660,6 +670,7 @@ public class GeneralData implements ProfileChangeListener {
         mLocalDriveBackupFreqDays = Helpers.parseInt(split, 70, 1);
         //mIsRemapFastForwardToSpeedToggleEnabled = Helpers.parseBoolean(split, 71, false);
         mIsRemapSToSpeedToggleEnabled = Helpers.parseBoolean(split, 72, true);
+        mIsConscryptEnabled = Helpers.parseBoolean(split, 73, false);
     }
 
     public void persistNow() {
@@ -685,7 +696,7 @@ public class GeneralData implements ProfileChangeListener {
                 mIsHideWatchedFromNotificationsEnabled, mChangelog, mPlayerExitShortcut, null, mIsFullscreenModeEnabled, null,
                 mIsRememberPinnedPositionEnabled, mSelectedItems, mIsFirstUseTooltipEnabled, mIsDeviceSpecificBackupEnabled, null,
                 null, mSearchExitShortcut, mGDriveBackupFreqDays, mLocalDriveBackupFreqDays, null,
-                mIsRemapSToSpeedToggleEnabled));
+                mIsRemapSToSpeedToggleEnabled, mIsConscryptEnabled));
     }
 
     @Override
